+++ src/tests/minesweeper_test.py
@@ -28,9 +28,11 @@ class TestMinesweeper(unittest.TestCase):
         self.game.reveal_cell(1,1)
         self.assertTrue(self.game.board[1][1]["isRevealed"])
 
-    def test_add_flag(self):
-        self.game.add_flag(1,1)
+    def test_toggle_flag(self):
+        self.game.toggle_flag(1,1)
         self.assertTrue(self.game.board[1][1]["isFlagged"])
+        self.game.toggle_flag(1,1)
+        self.assertFalse(self.game.board[1][1]["isFlagged"])
     
     def test_cell_with_no_neighbors(self):
         # 5x5 board with the center cell having no mine neighbors
@@ -38,7 +40,7 @@ class TestMinesweeper(unittest.TestCase):
         self.set_mines_for_testing(size=5, positions=positions)
 
         self.assertFalse(self.game.board[2][2]["isMine"])
-        self.assertFalse(self.game.board[2][2]["neighborCount"], 0)
+        self.assertEqual(self.game.board[2][2]["neighborCount"], 0)
 
     def test_cell_with_eight_neighbors(self):
         # 3x3 board with the center cell having eight
@@ -46,4 +48,32 @@ class TestMinesweeper(unittest.TestCase):
         self.set_mines_for_testing(size=3, positions=positions)
 
         self.assertFalse(self.game.board[1][1]["isMine"])
-        self.assertFalse(self.game.board[2][2]["neighborCount"], 8)
+        self.assertEqual(self.game.board[1][1]["neighborCount"], 8)
+    
+    def test_new_game(self):
+        # Tests that new_game initializes board correctly
+        game = Minesweeper()
+        game.new_game()
+        
+        self.assertEqual(len(game.board), game.size)
+        self.assertEqual(len(game.board[0]), game.size)
+
+        self.assertTrue(all(cell["isRevealed"] == False for row in game.board for cell in row))
+    
+    def test_reveal_adjacent_cells(self):
+        # Tests that revealing a cell with no adjacent mines reveal other with none
+        positions = [(2,0), (2,1), (2,2)]
+        self.set_mines_for_testing(size=3, positions=positions)
+        self.game.reveal_cell(0,0)
+
+        self.assertTrue(self.game.board[0][0]["isRevealed"])
+        self.assertTrue(self.game.board[0][1]["isRevealed"])
+        self.assertTrue(self.game.board[0][2]["isRevealed"])
+        self.assertTrue(self.game.board[1][0]["isRevealed"])
+        self.assertTrue(self.game.board[1][1]["isRevealed"])
+        self.assertTrue(self.game.board[1][2]["isRevealed"])
+        self.assertFalse(self.game.board[2][0]["isRevealed"])
+        self.assertFalse(self.game.board[2][1]["isRevealed"])
+        self.assertFalse(self.game.board[2][2]["isRevealed"])
+
+