+++ documentation/changelog.md
@@ -13,3 +13,10 @@
 - Added a text based display as a placeholder
 - Tested revealing and adding flags to cells
 - Tested that adjacent mines are counted correctly
+
+## Week 5
+- Created UI for the game, allowing play with a mouse
+- Added functionality of toggling flags
+- Revealing cells now reveals other non-mine adjacent cells
+- Added a game over property if the user clicks on a mine
+- Added tests for revealing cells and new_game